python/requirements.txt
# Phase 1: 基础框架与菜单系统
# 更新日期: 2026-02-11

# ============ 配置管理 ============
PyYAML==6.0.1              # YAML 配置文件读写

# ============ 交互界面 ============
questionary==2.0.1         # 交互式菜单
rich==13.7.0               # 终端美化（表格、进度条、面板）

# ============ 命令行框架 ============
click==8.1.7               # CLI 框架

# ============ 工具库 ============
python-dateutil==2.8.2     # 日期处理

# ============ Phase 2 将添加 ============
# playwright==1.42.0       # 网页自动化
# aiohttp==3.9.1           # 异步 HTTP
# beautifulsoup4==4.12.3   # HTML 解析
# tqdm==4.66.1             # 进度条

# ============ Phase 4 将添加 ============
# pandas==2.2.0            # 数据处理
# numpy==1.26.3            # 数值计算
# matplotlib==3.8.2        # 可视化
# seaborn==0.13.1          # 高级可视化
# jieba==0.42.1            # 中文分词
# wordcloud==1.9.3         # 词云生成
# Pillow==10.2.0           # 图像处理
# jinja2==3.1.3            # HTML 模板
